chore: refactor types out of atoms
