+++ web/containers/Providers/Jotai.tsx
@@ -4,6 +4,8 @@ import { PropsWithChildren } from 'react'
 
 import { Provider, atom } from 'jotai'
 
+import { FileInfo } from '@/types/file'
+
 export const editPromptAtom = atom<string>('')
 export const currentPromptAtom = atom<string>('')
 export const fileUploadAtom = atom<FileInfo[]>([])
@@ -14,11 +16,4 @@ export const selectedTextAtom = atom('')
 
 export default function JotaiWrapper({ children }: PropsWithChildren) {
   return <Provider>{children}</Provider>
-}
-
-export type FileType = 'image' | 'pdf'
-
-export type FileInfo = {
-  file: File
-  type: FileType
 }
