+++ web/types/file.d.ts
@@ -1,0 +1,6 @@
+export type FileType = 'image' | 'pdf'
+
+export type FileInfo = {
+  file: File
+  type: FileType
+}
